move checkpointing stuff to mlperf helpers
